Renamed the entities folder to model
