+++ trackr-app/app/src/main/java/com/trackr/trackr_app/model/Event.kt
@@ -1,4 +1,4 @@
-package com.trackr.trackr_app.entities
+package com.trackr.trackr_app.model
 
 import androidx.room.*
 import java.time.LocalDate
